http_caldav_sched.c:propcmp() compare the DATE-TIME properties by absolute value
The effect is that changing the DTSTART timezone and value do not trigger iTIP message,
if the start/end/due times converted to UTC do not change.

Comparing only the value with icalproperty_get_value_as_string() ignores
the TZ information and does not trigger a change, when the timezone changes,
the relative value remains the same, but the value converted to UTC does change.