Makefile.am: do not distribute the _err.c and _err.h files
Build imap_err.c only when AM conditional SERVER is on, as it is needed only
for libcyrus_imap, and the latter is build only when SERVER is on.

Build nntp_err.c only when AM conditional NNTPD is on.

Building x_err.c, builds x_err.h, so there is no need to request building
x_err.h explicitly.