mboxname: uniquer mboxname_todeleted suffixes
If the same mboxname gets deleted twice in the same second (i.e.
delete-create-delete) the latter would get the same new name
and be refused (NO).

This is not a real problem, but imaptest runs fast enough to
tickle it and the false positives are annoying.

Mixing the timestamp's microseconds into the existing suffix
(rather than appending or something else) to avoid breaking
middleware that wants to pattern match these filenames.