configure.ac: don't auto-enable cunit tests
Our cunit tests do not currently work in a VPATH build due to design
of cunit.pl, so we can't successfully run make distcheck when they
auto-enable.

We need to be able to run make distcheck much more than we need it to
automatically run the unit tests! So until cunit.pl is vpath-friendly,
you need to explicitly configure for it.