testrunner.pl: skip slow tests by default, --slow to override
Name your test function "test_whatever_slow" for it to be affected by
this feature.  Rule of thumb: if it takes like a minute or more to
run, it probably wants this.

We want Cassandane to be fast so we can run it frequently, but we also
want it to be thorough, and this is a convenient way to achieve both
goals!

If you explicitly request a slow test by naming it on the command line,
this will also override the "skip slow tests by default" behaviour.

Closes #80