caldav_sched: use ALL known addresses of the principal for scheduling purposes
This replaces the previous behaviour of "only schedule as the first address / aka
either the one requested by header or the first on the calendar-user-address-se"
with the much more sensible "treat every address on that set as local, and don't
schedule to it".

This MIGHT send multiple updates if you update multiple attendees and they're all
you.  It'll write one email per address that you happen to own.

What it doesn't do is check if you've fiddled anything else, but that should really
never happen anyway.