caldav: reverse the address-set over the wire
This removes the "preferred=1" since we won't need it any more,
and it reverses the strarray stored in Cyrus such that it's got the
LAST value first in the internal storage format.

This allows for the first value to be the default internally,
but for the wire format to match how Apple and hence everyone
else seems to be treating it.