imapoptions: add last modified field
Versions obtained from git history.  Tags prior to 2.3.17 are not
ancestors of master for some reason (rebase push between 2.3.16 and
2.3.17?) so anything older than 2.3.17 looks like it first appeared
in 2.3.17.

The last modified field is either the version in which the option
was first released, or the version in which the option's default
value, accepted values, or behaviour changed.  (You don't need to
bump the version if you're just correcting bad documentation, for
example.)