twoskip: add padding to ensure all pointers and crc32head fit within the same block
These are the parts of a record which are rewritten during later changes, so they
must be within the same 512 bytes to get clean crash recovery.  The earlier commits
in this series made it safe to read these blanks back out during dump and recovery2
which are the only things which care about the space between records.  These databases
can be read safely from old databases regardless of the blanks, but will fail on
recovery and dump.