Instance: check have_syslog_replacement more thoroughly
utils/syslog_probe is run with the same environment as any other
cyrus binary (including LD_PRELOAD and CASSANDANE_SYSLOG_FNAME).

If we can't find the magic word that it syslogs, we won't be able
to find anything else we're looking for either (probably the
LD_PRELOAD of utils/syslog.so has been denied), so we simply disable
checks for errors in syslog.