jmap_calendar: don't duplicate ACL from parent entry on create
The parent contains 'lr' for everybody who has share access into the user,
and if they get read access to the sub calendar they will be notified!

Either way - the share.With will change the ACLs to what we actually
want later - but better for them to never be set than set temporarily.