Remove SSE4.2 dependency checks and related code
SSE4.2 was checked to use a hardware native x86 processors instruction
for CRC32c calculation. Nevertheless, its result was not compatible with
the existing CRC32 algorithm, so it was never actually used anywhere in
the project, but the code was still getting compiled introducing
distribution issues for those processors that had no such instruction.

Actually, there probably was no such issue in practice as there was a
check at runtime in crc32c_init() for availability of the instruction,
though it might have had some issues in some corner cases on some
platforms and it was confusing for the maintainers.

In the future if a need arises to revive this functionality, just undone
this commit.

Closes #3102 (2/2)