Clarifications specialusealways: 1
As of RFC 6154 Section 2 an IMAP server that supports the SPECIAL-USE extension
MAY include any or all of the following attributes in responses to the
non-extended IMAP LIST command…

The option specialusealways alters exactly the behaviour of the unextended
LIST command.  Describing the behaviour of specialusealways in terms of the
undocumented XLIST command is suboptimal.

https://github.com/cyrusimap/cyrus-imapd/issues/2410