jmap_mailbox: change 'parent_id' handling to match external
Instead of toplevel folders having parent_id of Inbox, we use flags
on the entry and have a null parent_id for toplevel folders.  This
aligns them exactly with the external world and means that we're
not conflating them with subfolders of user.foo.INBOX.

There was also some case confusion in here for error handling cases.
This design is simpler and more consistent.