Fix faulty cyrusdb_foreach() call in mboxlist_sublist which can lead to assertion failures.
If the subscription database is twoskip, mboxlist_sublist() results in an
assertion failure in myforeach() in lib/cyrusdb_twoskip.c.

Parameters four and five to cyrusdb_foreach() are callback pointers, but only
parameter four may be null; the twoskip myforeach() asserts this is true.

Swap arguments four and five in the cyrusdb_foreach() call in
mboxlist_sublist() so we pass the callback where it is required.