http_ca(l,rd)dav: list collections: html escape the displayname
Otherwise >, <, and & in the displayname of calendars/address
books are not converted to &gt; &lt; or &amp; and the
generated HTML might not work.

When the DAV:displayname contains ' or ", generating valid HTML for
  <button onclick='anyFunction("THE-DISPLAYNAME")'>B</button>
is very hard.  Therefore the rendered HTML in list_calendars()
includes an id attribute to the <td> with the displayname and
passes around the id.  The callee uses then
  window.confirm(document.getElementById(id).innerText)
which can show ' and " to the mortals.