imap/spool.c: don't needlessly set/free parameters
In parseheader() the value of the parameters headname, contents and
rawvalue is never read, so there is no need to assign values to the
arguments name, body and raw before invoking parseheader().

Before parseheader() returns, it either sets newly allocated memory in the
parameters headname, contents and rawvalue and the function returns zero, or it
assigns NULL to headname, contents and rawvalue.  This means, after
parseheader() returns and the parameter headname⇔argument name is falsy, then
the parameters contents and rawvalue ⇔ arguments body and raw would also be
NULL, so there is no point to call free() on them.

The above is shortened even further: when the parameter headname / argument name
is set to NULL, there is no point to set NULL also to contents and rawvalue, as
in this case after parseheader() returns, the values of the parameter body and
raw are not read.