Typos
Insisting on “DCO” imposes formalities, that serve self-purpose.  One cannot
assume that the committer understands English, or that the submitter has time or
will to read texts about symbolism in software contributions.  If the system wants
to see the text

  nrEAUIEUAIe eEEAIUEAUEAEUAI eauaieeuai »ℓ§444.3.72b)°»°ℓ§euaieauuae

in each commit, people will write this, or any other text, that the system wants to
see.  All such text, which presense is mandated by the system, has the same value.

Signed-off-by: Дилян Палаузов <git-dpa@aegee.org>