jmap_calendar: rewrite CalDAV ATTACH URLs to BINARY for iTIP
Calendar events may contain ATTACH properties where the URI value points
to a managed CalDAV attachment [1]. These attachments are not publicly
accessible, and recipients of the scheduling message might not be able
to open that attachment.

To mitigate, Cyrus rewrites such ATTACH properties to contain the attachment
data as BINARY value when sending iTIP. When receiving iTIP message, it
attempts to rewrite any BINARY ATTACH properties to managed attachments.

[1] https://tools.ietf.org/id/draft-ietf-calext-caldav-attachments