imapd: don't check index if status on a different mailbox
See: https://bugzilla.mozilla.org/show_bug.cgi?id=1708981

The likelyhood that a client is waiting on this data and
won't issue a NOOP or otherwise check for it is much lower
than the likelyhood that the client won't think to look in
the status response to a different mailbox for an update to
the index state.

If we are doing a STATUS on the selected mailbox, we still
need to do the `index_check` though, because that will make
sure the data is up-to-date.