FastMail: mark test_imap_list_notes as :min_version_3_1
It would probably work fine on 3.0, but new() now puts 'jmap' into
httpmodules.  On 3.0 and earlier (which do not have any JMAP at all),
this causes the master daemon to refuse to start.

Since this suite is supposed to replicate the configuration at
Fastmail, which is always somewhere close to Cyrus master, it didn't
seem worth the trouble to conditionalise the config by Cyrus version.