sync_support: also sync all the historical names for each mailbox!
This lets us detect renames on the master and apply them to the replica
even without a RENAME logline, which means that we will no longer get
the case where a split of sync/log data could cause sync_client to try
to sync just one of the names and not realise that it was a renamed folder
and should be a rename rather than delete and copy.

This also broke in the past, because you could try to copy the new name
through and get a "uniqueid already exists" error from the replica.

Signed-off-by: Bron Gondwana <brong@fastmail.fm>