annotate: store values as dlists rather than binary
This is so much easier to reason about and parse.

The previous format had many moons of cruft in it, and isn't
even upgrade safe because it hard codes internal binary
lengths, so could be 4 or 8 bytes for an 'unsigned long'
depending on what compiler was used, and would fail to
parse when read with a different compiler.