sync_support: always sort annotations
This is needed because the [.OwNeR.] magic key now sorts differently
than the alphabetical place of the owner, so an annotations database
sort won't always be the same at both ends with different versions.

Rather than hoping it stays stable, it's safer to just always sort.
By doing it in the sync_annot_list, it's always sorted both off the
wire and when reading from the databases.