message: add function to extract conversations for a given message
When a message is added to a mailbox, it is determined if this message
belongs to an already existing conversation. Its conversation id is
then set accordingly. Up until now this only was possible for messages
that already got an index record. This patch changes the internal
conversation lookup code to also determine existing conversations
for a message that only exists in-memory.