jmap_backup: don't take the per-method conversations lock
Unfortunately, that lock means that all the other batching is pointless,
because there's an exclusive lock held on the whole user for the whole
period of the method call.

We want that for regular JMAP commands so that the state change is
contained and exact, but we don't want it for restore because restore
is arbitrarily slow.