conversations: when recalculating, upgrade to byid
This works by wiping everything if it's currently "byname", zeroing out
the folder entries so they get recalculated.

This MIGHT cause a one-off reverting of highestmodseq for folders and
messages if the last thing that happened to them was a purging of a
record, but that's OK for a one-off.