Instance: fix run_dbcommand_cb to handle literals without plus
It was removing the plus twice: once by assuming there was one,
and substr'ing it out, and then again by using a pattern match.

But if there was one, the pattern match was redundant; and if there
wasn't one, the substr shenanigans would lose the last byte of the
length instead, and break.