add a note about the incomplete and unportable libwrap test
On systems that wisely detect undefined references to symbols during
link time (including for dynamic-linked executables), this existing
libwrap detection macro for autoconf will fail with errors about
undefined references to global variables that libwrap callers are
expected to define.

Note this is just a note about the problem with this configure test.

The work-around for such systems, and where libwrap is used, is to set
ac_cv_lib_wrap_request_init=yes in the environment of the configure
script such that the test is assumed to have succeeded.

The real fix requires writing a more complex autoconf macro since
AC_CHECK_LIB() is not sophisticated enough to allow the extra code
needed to be included and something like AC_TRY_LINK() must be used
instead.  Note also that on some systems librwap also requires libnsl.
See for example the test in net-snmp's configure script.