jmap_mail: include Trash unseen counts in threadKeyword search
Up until now, evaluation of the threadKeyword query criteria
did not take the \Seen state of messages in the "Trash"
mailbox into account. The JMAP specification does not define
this special treatment for the "Trash" mailbox.

This patch updates threadKeyword search to include the \Seen
of messages within Trash. To preserve former behaviour,
queries must explicitly exclude the "Trash" mailbox, using
either an "inMailboxOtherThan" or a negated "inMailbox" criteria.
For such queries, the \Seen count of messages with "Trash"
is ignored.

Note that this patch only applies to the hard-coded "Trash"
mailbox name. Mailboxes with other name and role=trash always
counted towards the total \Seen count of a thread.