dav_db:CMD_DBUPGRADEv10 — DROP INDEX IF EXISTS idx_res_uid
Since d35f0dac07904bef the index is not created for new setups.  The index
was however not removed from older setups.

This change does not help for upgrades to 3.4, which upgrades were performed
before this change.  For such setups, that is if you already run 3.4, the command

for x in $(find /var/imap -name '*.dav'); do sqlite3 $x "DROP INDEX IF EXISTS idx_res_uid"; done

executed under the 'cyrus' user, does perform the change, where '/var/imap' is
the configdirectory in imapd.conf .