sieve_db: use dav_close rather than sqldb_close
This fixes a crasher when running dav_reconstruct.  The problem is
that the mailbox_sieve_commit tries to close the DB, which is still
in a transaction and asserts.

dav_close is this:

```
EXPORTED int dav_close(sqldb_t **dbp)
{
    if (reconstruct_db) return 0;

    return sqldb_close(dbp);
}
```

Which precisely avoids this issue!  The other users of dav_db (caldav_db,
carddav_db) corretly use dav_close() in their close functions.