TestUrl - Do not mess with Cassandane's END/DESTROY blocks
We do this by creating an object which will call POSIX::_exit()
when it goes out of scope, and tucking that into our app wrapper
which runs in the fork.

POSIX::_exit() ends Perl without running END/DESTROY blocks.

Any way (that I've tested) that Perl attempts to end in that
child, first destroys our scope guard. This happens in the *run*
phase, which means other objects have not been cleaned up yet
in some indeterminate order (which happens in the destruction
phase).

Signed-off-by: Matthew Horsfall <wolfsage@gmail.com>