caldav_db: update jscal_objs table records also if 'alive' changed
If Cyrus replicates a calendar event, if first marks the existing
ical_objs entry as expunged, setting alive to 0. This alive status
propagates into the jscal_objs table.

Later on, the jscal_objs handling code only updates an existing
jscal_objs entry, if the contents of the iCalendar data changed.
This caused replication to leave standalone instance entries in
that table as not alive, when on the main store they are alive.