jmap_ical: return all standalone recurrence instances, not only first
The previous iCalendar to JMAP conversion assumed that there always
is a main component present, and if not it promoted the first
standalone instance in the VCALENDAR to the main component.

This is not an issue for the CalendarEvent/get codepaths, as they
preprocess the iCalendar component to contain at most one
standalone instance. But it this assumption does not hold for
arbitrary iCalendar data, such as found in email attachments.

Signed-off-by: Robert Stepanek <rsto@fastmailteam.com>