jmap_calendar: only read on-disk file once for each iCalendar data
Before JMAP Calendars, the getcalendarevents callback was called
exactly once for each iCalendar file. Now it gets called as many
times as there are standalone instances (or one main event) in the
data. For files with a large number of standalone instances, this
resulted in lots of calls of reading the on-disk data and index record.

This patch buffers the last read iCalendar data in the iterator,
and thus reduces the number of record_to_ical calls.

Signed-off-by: Robert Stepanek <rsto@fastmailteam.com>