configure.ac: don't permit 'timezone' as a variable name
There is a global of the same name in libc (time.h, which we always
include).  If we also use timezone as a local variable name, but
make a mistake, the compiler might not catch it.

So -- just don't allow 'timezone' as a variable name.

If you actually need to use the libc timezone global, you can
'#define REALLY_USE_LIBC_TIMEZONE' before including config.h to
expose it.