jmap_ical: allow creation time of override to differ from main event
Before, the 'created' timestamp of a recurrence override was not
allowed to differ from the main event. Now, the created property
can be set for new overrides, but the value must not be changed
later. As for main events, a created timestamp in the future is
clamped to server time.

Signed-off-by: Robert Stepanek <rsto@fastmailteam.com>