conversations: count email as unseen if unseen in any folder
Before, conversation folder counts only counted an email as unseen
if it was unseen in that folder. This conflicts with JMAP because
an Email object is reported as unseen if it unseen in any mailbox
where a copy of that email exists in.

Since folder unseen counts only are used for JMAP, this patch
changes the unseen count to implement the JMAP Email semantics.
Only the Trash folder counts unseen as previously.

Signed-off-by: Robert Stepanek <rsto@fastmailteam.com>