jmap_calendar: reset iCalendar iterator state for each calendar
This applies to CalendarEvent/get:

The iCalendar iterator state keeps the current iCalendar data
in-memory as long as the IMAP uid of the event resource remains
unchanged. But the iterator state was not reset when the iterator
switches to the next calendar mailbox.

Consequently, if the last IMAP uid of the previously processed calendar
matched the first IMAP uid in the new calendar, then CalendarEvent/get
erroneously returned the contents of the former event with only the
calendarIds and iCalendar uid having the correct value.

This patch also bumps the cache version of the JMAP CalendarEvent
cache, so that any cached bogus events are flushed from it.

Signed-off-by: Robert Stepanek <rsto@fastmailteam.com>