http_jwt: OpenSSL 3 compat EVP_PKEY_base_id
The WARNING section of the EVP_PKEY_base_id manpage in OpenSSL 3 says
that EVP_PKEY_base_id is "only reliable with EVP_PKEYs that have been
assigned an internal key with EVP_PKEY_assign_*", and the usage here of
base_id seems to be without EVP_PKEY_assign usage.

This same warning points to EVP_PKEY_is_a, which seems fine for this
purpose but is part of OpenSSL 3.

Signed-off-by: Dan Bungert <daniel.bungert@canonical.com>