ci: add github workflow
cyrus-imapd has been relying on Travis CI for builds. And Travis CI
has been pretty unreliable for our use-case as we keep running out of
'free OSS credits'.
This commit adds support for Github Actions for builds.

While most of what we did with Travis worked with Actions, a minor
change in the how cassandane stops processes was also needed and is a
part of this commit. The change was to use `stop_command()` method
instead of `_stop_pid()` method directly, making sure that a
`waitpid()` call is made after a `kill()`.

Signed-off-by: Partha Susarla <mail@spartha.org>