jmap_util: charset detection must not decrease printable characters
In some cases, the charset detection library might erroneously
guess a text to be encoded in a single byte encoding such as
ISO-8859-2. The resulting UTF-8 string then trivially contains no
invalid or replacement byte sequences but a large number of
single byte control characters.

To prevent this, this patch adds the heuristic that the amount of
printable characters must not decrease in relation to the total
number of all valid characters.

Signed-off-by: Robert Stepanek <rsto@fastmailteam.com>