search_expr: Remove trivial nodes during the normalisation process
When normalizing a search expression, the normalizing process can
generate too many nodes for the complexity limit while in fact the
normalized expression is quite simple. To overcome this, we reduce
the nnodes complexity when a node is freed, and if too many nodes are
created during a pass of the normalization process, a cleaning pass is
performed on the search tree.

Signed-off-by: Mildred Ki'Lya <mildred-pub.git@mildred.fr>