[jmap] Improve query of the NOT operator
The JMAP NOT operator was translated to a NOT(OR(..., ...)) search
expression instead of a AND(NOT(...), NOT(...)). The process of
normalizing search expressions forces the second form. Providing the
search expression already in the correct order removes complexity from
the normalization process, and can help not reaching the complexity
limit.

Signed-off-by: Mildred Ki'Lya <mildred-pub.git@mildred.fr>