assert: assert() must be an expression, not a statement
Note that while this is now syntactically compatible with the
C standard assert macro(), it is still not semantically compatible.

Specifically, the C standard assert() macro can be compiled out
entirely by defining NDEBUG during compilation, and thus is a
debugging aid only.

Ours cannot be compiled out, and therefore can be relied upon as
a run-time safety -- which we do, extensively!