cyrus_icalrestriction_check() strip all X-LIC-ERROR properties from the input
When the CUA PUTs an iCalendar object with X-LIC-ERROR property,
Cyrus IMAP rejects the upload, as the iCalendar object is assumed
to contain errors.  It does however not contain errors.

This can happen, when the CUA uses libical compiled with
ICAL_ALLOW_EMPTY_PROPERTIES=OFF and the CUA receives an iCalendar
object (per email) with LOCATION: property without value.  libical
transforms then the empty LOCATION: property in X-LIC-ERROR property.