ptrarray: do not memset buffer that gets free()d anyway
This fixes a potential issue highlighted by @dilyanpalauzov in
https://github.com/cyrusimap/cyrus-imapd/pull/4206

ptrarray_fini calls memset to set the ptrarray data to zero,
irrespective if it points to NULL or an allocated buffer.
This is safe under normal conditions, as a ptrarray with null
data is expected to have a member count. It is unsafe if
this condition does not hold.

This patch removes the call to memset altogether, similar to
how fini is implemented for strarray.

Signed-off-by: Robert Stepanek <rsto@fastmailteam.com>