libconfig: add BYTESIZE option type + config_getbytesize()
Allows byte sizes to be specified with units, e.g. "10KB" etc.

config_getbytesize() takes a second argument for the default unit
to assume if the value doesn't have one, so legacy config files
can be made to keep working if this is called with the historical
unit for that option.