itip: remove CLASS property from iTIP messages
This patch removes the iCalendar CLASS property when receiving and
sending iTIP messages. For JMAP Calendars, Cyrus enforces the access
classification of the CLASS property for sharees of a calendar mailbox.
It should be up to the calendar owner, not the ORGANIZER of an event,
to set the desired access level.

Signed-off-by: Robert Stepanek <rsto@fastmailteam.com>