jmap_admin: add JMAP admin extension for Cyrus admin tasks
This adds the non-standard JMAP admin extension to run
administrative tasks via JMAP.

As a first such method, the Admin/rewriteCalendarEventPrivacy
method rewrites the caldav database for calendar events where
the JMAP privacy property is set. This is to migrate existing
calendar events that had their privacy set due to the iCalendar
CLASS property but now need to determine it by the newly
introduced X-JMAP-PRIVACY property.

Signed-off-by: Robert Stepanek <rsto@fastmailteam.com>