tools/compile_st.pl: strip 'Command-line' from gperf output
We version control the output file from this command, and having
the name of the tempfile used to generate it embedded in the output
made it show up as different every time it was regenerated, even
though usually the only difference was the name of the tempfile.

Now it will only show up as different when something interesting
changed.