caldav_alarm: do not zero out nextcheck if sharee has no alarms
Cyrus processes calendar alarms in shared calendars by first
checking the calendar owner per-user data for alarms, then
for each sharee. It then stores the smallest, future UTC time
when to check alarms again for the event, where the zero value
disables any future alarms.

A bug in this logic caused Cyrus to not fire any alarms for
shared calendar events, if the calendar owner had alarms set
but none of the sharees.

Signed-off-by: Robert Stepanek <rsto@fastmailteam.com>