jmap_api: initialize jmap_changes struct to zero
Initialization of the jmap_changes struct is left to the
handler function first calling jmap_changes_parse. But
in at least one occasion this could get skipped with an
early return. Since the compiler did not report this
during compilation let's better make sure we zero the
structs memory properly.

Signed-off-by: Robert Stepanek <rsto@fastmailteam.com>