jmap_api: refactor JMAP state API to jmap_modseq function
This reduces the current JMAP state handling API to the single function
jmap_modseq. The former code was the result of multiple refactorings
during the initial implementation of JMAP. The result were functions that
handled state as opaque JSON strings, C string or as modseqs. This patch
changes that to always handle JMAP state as modseq.

Signed-off-by: Robert Stepanek <rsto@fastmailteam.com>