http_caldav: allow CalDAV alarms on calendar home (3.8)
This allows setting CalDAV alarm DAV properties on the
calendar home. Version 3.7 erroneously broke this and
a complete rewrite of default alarm support in Cyrus is
in the works.

In order not to include the current broken state in the
upcoming 3.8 release, this patch falls back to treating
CalDAV alarm DAV properties like any other "dead" DAV
property. This is consistent with version 3.6 and earlier.

Signed-off-by: Robert Stepanek <rsto@fastmailteam.com>