jmap_ical: do not set deprecated "timeZones" property
RFC 8984 allows the "timeZones" property to be set in a
JSCalendar event, but this is deprecated in jscalendarbis.

This patch refactors the "timeZones" property to the
Cyrus-internal "cyrusimap.org:timeZones" extension
property. This property is only meant for Cyrus-internal
use and currently is required to preserve Microsoft
VTIMEZONE definition in the iCalendar representation of
the event.

Signed-off-by: Robert Stepanek <rsto@fastmailteam.com>