jmap_ical: ignore JMAP timezone id in non-IANA VTIMEZONE
Cyrus preserved non-IANA timezone identifiers in the VTIMEZONE
representation of a custom JMAP timezone. Up until now,
Cyrus did not attempt to guess an IANA timezone for such
timezone definitions, causing JMAP clients to not being able
to determine the event start time.

This patch makes Cyrus to ignore JMAP timezone ids in existing
VTIMEZONE definitions and not preserve them in new VTIMEZONEs.
Consequently, it always attempts to guess an IANA timezone for
a non-IANA VTIMEZONE, unless this feature is disabled by setting
the no_guess flag in the converter.

Signed-off-by: Robert Stepanek <rsto@fastmailteam.com>