global.c: leniently support message-ids without brackets
RFC 5322 is clear: A valid message-id must be enclosed
in "<" and ">" brackets.

Unfortunately, some email implementations do not get
this right and dump message-ids without brackets in
their Message-Id and References headers.

This patch updates the find_msgid function to accept
such malformed message-ids iff they are of the form

    WSP dot-atom-text '@' dot-atom-text [WSP|EOF]

and return them with proper closing and end brackets.

This does not apply to the In-Reply-To header, because
broken MUAs tend to dump email addresses in that header.

Signed-off-by: Robert Stepanek <rsto@fastmailteam.com>