charset.testc: add tests for handling of '/' and '>' in HTML
This adds regression tests for Cyrus current handling of
the characters '/' and '>' during HTML extraction.
The current implementation treats any occurence of these
characters as the definite end of a XML tag ('>') or a
probable end ('/'). This often is the right thing to do,
but is wrong in the case of quoted attribute values
containing such characters.

The purpose of the tests is catch regressions should we ever
change the XML parser logic.

Signed-off-by: Robert Stepanek <rsto@fastmailteam.com>