charset: extract URLs from and "alt" text from HTML
This updates HTML to plain text extraction to preserve
the values of the HTML anchor and area tag "href"
attributes, as well as the HTML img tag "src" and "alt"
attributes.

The value of the "href" and "src" attributes is enclosed
in angle brackets. It is emitted at the end of the HTML
tag, such as:

     <a href="https://example.com">Click here</a>

     Click here <https://example.com>

The value of the img "alt" attribute is surrounded by
round brackets and follows the value of the "src"
attribute, separated by a single space. An "alt"
attribute without a "src" attribute is ignored.
For example:

     <img alt="A cat" src="cat.png"/>

     <cat.png> (A cat)

This change only applies to attribute values that are
enclosed either by double or single quotes. Unquoted
attribute values remain ignored.

Signed-off-by: Robert Stepanek <rsto@fastmailteam.com>