testrunner.pl: add --slow-only option
We already had --slow, to run everything including the _slow tests,
which are otherwise skipped.

Now we have --slow-only, to run only the _slow tests

Use this after you've done a regular run, and then remember that you
want to run the slow tests too, but don't want to sit through another
whole run of the regular tests too.

If you name non-slow tests on the command line, but also specify
the --slow-only option, it'll warn you and run the non-slow tests.