github actions: set safe.directory for build directory
actions/checkout@v3 does this temporarily for itself, but removes
it when it's done.

Even though we're no longer fetching the tags manually, we still need git
commands to work (in particular, tools/git-version.sh used by the build
needs "git describe" to set the version string), so we need to set
safe.directory manually for that.

While we're here, log the result of tools/git-version.sh so if something
is gonna break due to a weird version, the autopsy has that information
up front.