jmap_ical: remove experimental Alert.uid property
This only got introduced during development and was meant
to ensure using sane UIDs in VALARMs. However, the released
Cyrus code stores the JMAP id of the Alert object as the
VALARM UID. This makes converting snoozed alarms simpler,
as there is no risk of broken relatedTo references. The
drawback is that it's up to the client to provide a sane UID
value.

For now, we'll keep using what's already released and might
deal with separating JMAP Alert ids and VALARM UIDs later.

Signed-off-by: Robert Stepanek <rsto@fastmailteam.com>