doc generation: stop using custom manpage builder
Why is the current one broken with newer libraries?

I don't know!  But we also no longer know why we had a customer builder,
either.  At first glance, the output without the custom code seems fine.
We should inspect it more, and then we can decide if we can go forward
with this and also delete the custom code!