Makefile.am: make 'check' etc fail on cunit valgrind errors
This makes it so that cunit-under-valgrind will fail if valgrind
detects memory errors, instead of reporting the valgrind errors
to the terminal but still succeeding.  This will be more useful
for automated testing.

Specifically, this affects the following make targets:
* check
* check-local
* check-%

The check-discrete make target already behaved like this, and is
no longer a special case.