conversations: sort thread by createdmodseq if same internaldate
This gives us a stable sort when multiple external messages are appended
very quickly, giving them the same internaldate.

This is being added without a new version number, as it extends a data
structure in a backwards and forwards compatible way.  If not set, the
CREATEDMODSEQ field will be zero, which is fine.  This will still lead
to correct sorting for newly added emails on any thread.  The only
slight weirdness might be if an existing message in a thread gets moved
or copied into a new folder, which will lead to it getting a
createdmodseq value, and hence sorting to the end for that internaldate.
This will be very rare, and a convdb rebuild will fix it.